Ensure that returned value includes specified name
Prior to this commit, invoking the `AccountPolicy` resource with the `Get` method would always return `$Null` for the Name property; though this property is itself never used by the resource, integrating tools are unable to tell as the Name property is also the `Key` (as defined in the MOF) for this resource; meaning when comparing the existing state of the resource it will always cause a mismatch between the specified Name and the "actual" representation of the resource on the machine, as reported by the `Get-TargetResource` function.

This commit modifies the `Get-TargetResource` function for the AccountPolicy DSC resource to _always_ return the specified Name for the resource in order to prevent churn in secondary tools such as Puppet, Ansible, or Chef.